% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_sensitivity.R
\name{ae_sensitivity}
\alias{ae_sensitivity}
\title{Sensitivity analysis for Angrist and Evans (1998)}
\usage{
ae_sensitivity(savedir = getwd())
}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Evaluate the sensitivity of Angrist and Evans (1998) to the
monotonicity condition then create a figure illustrating it.

Bias of Wald for LATE is:
(Pr(DF)/(Pr(DF) + Pr(CP))) * (LATE - DATE)
where
DATE = defier average treatment effect
}
