% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_monotonicity.R
\name{average_monotonicity}
\alias{average_monotonicity}
\title{Illustrate average monotonicity}
\usage{
average_monotonicity(savedir = getwd())
}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Illustrate average monotonicity
}
