% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card}
\alias{card}
\title{Card (1993/1995) data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3010 rows and 34 columns.
}
\usage{
card
}
\description{
Classic NLSY extract used in Card (1993).
}
\keyword{datasets}
