% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_app}
\alias{card_app}
\title{Empirical results for the Card illustration}
\usage{
card_app(savedir = getwd(), nbs = 2, nsplits = 1)
}
\arguments{
\item{savedir}{Directory where results will be saved.}

\item{nbs}{Number of bootstrap replications for p-score weighting estimates}

\item{nsplits}{Number of split repetitions for DDML}
}
\description{
Empirical results for the Card illustration
}
