% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gelbach}
\alias{gelbach}
\title{Gelbach (2002) data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10932 rows and 23 columns.
}
\usage{
gelbach
}
\description{
This is the "single mothers" sample already subset to youngest child age 5.
It's the same sample as used in Gelbach Table 7, column 3.
}
\keyword{datasets}
