% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelbach.R
\name{gelbach_bounds}
\alias{gelbach_bounds}
\title{Bounds for the Gelbach illustration}
\usage{
gelbach_bounds(savedir = getwd())
}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Bounds for the Gelbach illustration
}
