% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelbach.R
\name{gelbach_comparisons}
\alias{gelbach_comparisons}
\title{Comparisons for the Gelbach illustration}
\usage{
gelbach_comparisons(savedir = getwd(), outcome = "work79")
}
\arguments{
\item{savedir}{Directory where results will be saved.}

\item{outcome}{A string for the name of the Y variable}
}
\description{
Comparisons for the Gelbach illustration
}
