% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivhandbookReplication-package.R
\docType{package}
\name{ivhandbookReplication-package}
\alias{ivhandbookReplication}
\alias{ivhandbookReplication-package}
\title{ivhandbookReplication: Replication of Mogstad and Torgovitsky (2024, Handbook of Labor Economics)}
\description{
Reproduces figures and tables that require some computation. Three data sets are included.
}
\author{
\strong{Maintainer}: Alexander Torgovitsky \email{atorgovitsky@gmail.com}

}
\keyword{internal}
