% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_weighting.R
\name{multi_weighting}
\alias{multi_weighting}
\title{Illustration of weights for IV estimands with multivalued instruments}
\usage{
multi_weighting(savedir = getwd())
}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Computing the IV estimand for different hypothetical experiments
Then create a plot to visualize the difference.
}
