% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{runall}
\alias{runall}
\title{Run everything (replicate)}
\usage{
runall(resultsdir = fs::path_wd("results"))
}
\arguments{
\item{resultsdir}{Directory where results will be saved. This is the
top-level directory. A timestamped subdirectory will be created within it,
and then results will be saved inside that subdirectory.}
}
\description{
Run everything (replicate)
}
