% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_params.R
\name{shared_savedir}
\alias{shared_savedir}
\title{Shared parameters}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Define parameters that are reused to avoid duplication
}
\keyword{internal}
