% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stevenson}
\alias{stevenson}
\title{Stevenson (2018) data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 331971 rows and 50 columns.
}
\source{
\url{https://github.com/scunning1975/mixtape/raw/master/judge_fe.dta}
}
\usage{
stevenson
}
\description{
Judge data from Stevenson (2018) via Cunningham ()
}
\keyword{datasets}
