% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unordered_treatments.R
\name{unordered_treatments}
\alias{unordered_treatments}
\title{Enumerate groups for unordered treatments, compare assumptions,
make a table}
\usage{
unordered_treatments(savedir = getwd())
}
\arguments{
\item{savedir}{Directory where results will be saved.}
}
\description{
Enumerate groups for unordered treatments, compare assumptions,
make a table
}
